refactor: reorganize controller folder structure
- Move product controllers to app/Http/Controllers/Product
- Move purchase controllers to app/Http/Controllers/Purchase
- Move sale/POS controllers to app/Http/Controllers/Sale
- Move user controllers to app/Http/Controllers/User
- Move payment controllers to app/Http/Controllers/Payment
- Move stock-related controllers to app/Http/Controllers/Stock
- Create separate DashboardController at root level
- Improve code organization and maintainability
