+++ app/Http/Controllers/DashboardController.php
@@ -1,9 +1,8 @@
 <?php
 
-namespace App\Http\Controllers\Dashboard;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
-use App\Services\DashboardService;
+use App\Services\General\DashboardService;
 use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
